+++ src/main/scala/coulomb/Quantity.scala
@@ -32,43 +32,6 @@ import define._
 import test._
 
 
-
-// assuming inputs (U, P) are already from some canonical sig; so a type U will never pre-exist in M
-trait InsertSortedUnitSig[U, P, M] {
-  type Out
-}
-object InsertSortedUnitSig {
-  type Aux[U, P, M, O] = InsertSortedUnitSig[U, P, M] { type Out = O }
-
-  implicit def evidence0[U, P]: Aux[U, P, HNil, (U, P) :: HNil] =
-    new InsertSortedUnitSig[U, P, HNil] { type Out = (U, P) :: HNil }
-
-  implicit def evidence1[U, P, U0, P0, MT <: HList](implicit lte: XIntGT.Aux[P, P0, False]): Aux[U, P, (U0, P0) :: MT, (U, P) :: (U0, P0) :: MT] =
-    new InsertSortedUnitSig[U, P, (U0, P0) :: MT] { type Out = (U, P) :: (U0, P0) :: MT }
-
-  implicit def evidence2[U, P, U0, P0, MT <: HList, O <: HList](implicit gt: XIntGT.Aux[P, P0, True], rc: Aux[U, P, MT, O]): Aux[U, P, (U0, P0) :: MT, (U0, P0) :: O] =
-    new InsertSortedUnitSig[U, P, (U0, P0) :: MT] { type Out = (U0, P0) :: O }
-
-}
-
-trait SortUnitSig[M, N, D] {
-  type OutN
-  type OutD
-}
-
-object SortUnitSig {
-  type Aux[M, N, D, ON, OD] = SortUnitSig[M, N, D] { type OutN = ON; type OutD = OD }
-
-  implicit def evidence0[N, D]: Aux[HNil, N, D, N, D] =
-    new SortUnitSig[HNil, N, D] { type OutN = N; type OutD = D }
-
-  implicit def evidence1[U, P, MT <: HList, N, D, NO, NF, DF](implicit pos: XIntGT.Aux[P, XInt0, True], ins: InsertSortedUnitSig.Aux[U, P, N, NO], rc: Aux[MT, NO, D, NF, DF]): Aux[(U, P) :: MT, N, D, NF, DF] =
-    new SortUnitSig[(U, P) :: MT, N, D] { type OutN = NF; type OutD = DF }
-
-  implicit def evidence2[U, P, MT <: HList, N, D, NP, DO, NF, DF](implicit neg: XIntLT.Aux[P, XInt0, True], n: XIntNeg.Aux[P, NP], ins: InsertSortedUnitSig.Aux[U, NP, D, DO], rc: Aux[MT, N, DO, NF, DF]): Aux[(U, P) :: MT, N, D, NF, DF] =
-    new SortUnitSig[(U, P) :: MT, N, D] { type OutN = NF; type OutD = DF }
-}
-
 trait SigToUnit[S] {
   type Out
 }
